refactor: move rbac files to subdirectory components
