+++ src/components/projects/rbac/policy.vue
@@ -1,0 +1,19 @@
+<template>
+  <div class="rbac-policy">
+    policy
+  </div>
+</template>
+
+<script>
+export default {
+  data () {
+    return {}
+  }
+}
+</script>
+
+<style lang="less" scoped>
+.rbac-policy {
+  background-color: inherit;
+}
+</style>
